refact in the result handling
