+++ DataStructures/DataStructures.csproj
@@ -53,6 +53,7 @@
     <Compile Include="ReferenceSample\ReferenceSample.cs" />
     <Compile Include="ReferenceSample\ReferenceValue.cs" />
     <Compile Include="ResultBase.cs" />
+    <Compile Include="ToolSpecifications\ConditionEvaluaitonResult.cs" />
     <Compile Include="ToolSpecifications\Quantity.cs" />
     <Compile Include="ToolSpecifications\RotatableToolSpecification.cs" />
     <Compile Include="ToolSpecifications\ConditionBase.cs" />
