+++ DataStructures/ResultBase.cs
@@ -4,7 +4,7 @@ using System.Xml.Linq;
 
 namespace DataStructures
 {
-    public abstract class ResultBase : IResult
+    public abstract class ResultBase : IResult, IFormattable
     {
         protected ResultBase(DateTime creationTime, bool successful)
         {
@@ -15,14 +15,15 @@ namespace DataStructures
         public virtual DateTime CreationTime { get; }
         public virtual bool Successful { get; }
 
-        public XElement SaveToXml(XElement input)
-        {
-            throw new NotImplementedException();
-        }
+        public abstract XElement SaveToXml(XElement input);
 
-        public bool LoadFromXml(XElement input)
+        public abstract bool LoadFromXml(XElement input);
+
+        public abstract string ToString(string format, IFormatProvider formatProvider);
+
+        public override string ToString()
         {
-            throw new NotImplementedException();
+            return $"CreationTime:{CreationTime}Successful{Environment.NewLine}{Successful}{Environment.NewLine}";
         }
     }
 }
