+++ Interfaces/Interfaces.csproj
@@ -37,12 +37,14 @@
     <Reference Include="System.Xml.Linq" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="BaseClasses\EventArgs.cs" />
+    <Compile Include="BaseClasses\Relations.cs" />
     <Compile Include="BaseClasses\Enums.cs" />
+    <Compile Include="BaseClasses\ToolNames.cs" />
     <Compile Include="Calculation\ICalculation.cs" />
     <Compile Include="Calculation\ICalculationContainer.cs" />
     <Compile Include="Calculation\ICalculationSettings.cs" />
     <Compile Include="DataAcquisition\IDataCollector.cs" />
-    <Compile Include="BaseClasses\Enumerations_BaseClasses.cs" />
     <Compile Include="Evaluation\IMathing.cs" />
     <Compile Include="MeasuredData\IGenericMeasurementPoint.cs" />
     <Compile Include="Misc\IComponent.cs" />
